+++ projects/igniteui-angular/src/lib/grids/summaries/summary-row.component.ts
@@ -107,6 +107,22 @@ export class IgxSummaryRowComponent implements DoCheck  {
         return this.grid.pinnedColumns;
     }
 
+
+    /**
+     * @hidden
+     */
+    public get pinnedStartColumns(): ColumnType[] {
+        return this.grid.pinnedStartColumns;
+    }
+
+
+    /**
+     * @hidden
+     */
+    public get pinnedEndColumns(): ColumnType[] {
+        return this.grid.pinnedEndColumns;
+    }
+
     /**
      * @hidden
      */
@@ -114,9 +130,10 @@ export class IgxSummaryRowComponent implements DoCheck  {
         return this.grid.unpinnedColumns;
     }
 
-    public getContext(row) {
+    public getContext(row, cols) {
         return {
-            $implicit: row
+            $implicit: row,
+            columns: cols
         };
     }